can get qr data
