+++ app/src/main/java/com/batch/dojo19tobecontinued/friendlist/FriendListFragment.kt
@@ -3,17 +3,16 @@ package com.batch.dojo19tobecontinued.friendlist
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.animation.OvershootInterpolator
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.batch.dojo19tobecontinued.R
-import com.batch.dojo19tobecontinued.friendlist.model.Friend
 import jp.wasabeef.recyclerview.adapters.AlphaInAnimationAdapter
 import jp.wasabeef.recyclerview.adapters.ScaleInAnimationAdapter
 import kotlinx.android.synthetic.main.fragment_friendlist.*
@@ -37,6 +36,7 @@ class FriendListFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        Log.d("DojoApp", "onViewCreatedだよ")
         viewModel.loadDatabase(requireContext())
 
         friend_recycler_view.apply {
@@ -49,16 +49,18 @@ class FriendListFragment : Fragment() {
                 setInterpolator(OvershootInterpolator(.100f))
             }
         }
-//        add_friend_button.setOnClickListener { viewModel.addFriendTest() }
 
         add_friend_button.setOnClickListener {
-            viewModel.openCamera(requireActivity())
+            try {
+                val intent = Intent("com.google.zxing.client.android.SCAN")
+                intent.putExtra("SCANMODE", "QR_CODE_MODE")
+                startActivityForResult(intent, 0)
+            } catch (e: Exception) {
+                val marketUri = Uri.parse("market://details?id=com.google.zxing.client.android")
+                val marketIntent = Intent(Intent.ACTION_VIEW, marketUri)
+                startActivity(marketIntent)
+            }
         }
-//        viewModel.allFriends.observe(this, Observer { friends ->
-//            friends?.let {
-//                friendListAdapter.setFriends(it)
-//            }
-//        })
 
         viewModel.state.observe(this, Observer {
             checkState(it)
@@ -66,16 +68,18 @@ class FriendListFragment : Fragment() {
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        Toast.makeText(requireContext(), "読み取ったよ", Toast.LENGTH_SHORT).show()
-        viewModel.getQRReadResult(requireContext(), requestCode, resultCode, data)
+        Log.d("DojoApp", "onActivityResultだよ")
+        viewModel.getQRReadResult(requestCode, resultCode, data)
     }
 
     private fun checkState(state: FriendListState) {
+        Log.d("DojoApp", "checkState呼ばれたよ．friendListの中身${state.friendList}")
         if (state.friendList != null) {
+            Log.d("DojoApp", "friendListAdapterにセットだよ")
             friendListAdapter.setFriends(state.friendList)
         }
         if (state.isReadSuccess && state.readResultData != null) {
+            Log.d("DojoApp", "saveFriendだよ")
             viewModel.saveFriend(state.readResultData)
         }
     }
