+++ app/src/main/java/com/batch/dojo19tobecontinued/friendlist/FriendListViewModel.kt
@@ -4,7 +4,7 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import android.widget.Toast
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -21,9 +21,9 @@ class FriendListViewModel : ViewModel() {
     private val _state = MutableLiveData(FriendListState.INITIAL)
     val state: LiveData<FriendListState> = _state
 
-    //    lateinit var allFriends: LiveData<List<Friend>>
-    private lateinit var repository: FriendRepository
+    lateinit var friendList: LiveData<List<Friend>>
 
+    private lateinit var repository: FriendRepository
     private val job = Job()
     private val coroutineContext: CoroutineContext
         get() = job + Dispatchers.Main
@@ -32,37 +32,12 @@ class FriendListViewModel : ViewModel() {
     fun loadDatabase(context: Context) {
         val friendDao = AppDatabase.getFriendDatabase(context).friendDao()
         repository = FriendRepository(friendDao)
+//        friendList = repository.allFriends
         _state.value = _state.value?.copy(friendList = repository.allFriends.value)
-    }
-//        allFriends = repository.allFriends
-
-//    fun addFriendTest() {
-//        val friend = Friend(
-//            fullName = "HOGE${_state.value?.friendList?.value?.size}",
-//            githubID = "@githubhoge",
-//            twitterID = "twitterhoge"
-//        )
-//        scope.launch(Dispatchers.IO) {
-//            repository.insert(friend)
-//        }
-//        _state.value = _state.value?.copy(friendList = repository.allFriends)
-//        allFriends = repository.allFriends
-//    }
-
-    fun openCamera(activity: Activity) {
-        try {
-            val intent = Intent("com.google.zxing.client.android.SCAN")
-            intent.putExtra("SCANMODE", "QR_CODE_MODE")
-            activity.startActivityForResult(intent, 0)
-        } catch (e: Exception) {
-            val marketUri = Uri.parse("market://details?id=com.google.zxing.client.android")
-            val marketIntent = Intent(Intent.ACTION_VIEW, marketUri)
-            activity.startActivity(marketIntent)
-        }
     }
 
     fun saveFriend(data: Intent) {
-        val content = data.getStringExtra("SCAN_RESLT")
+        val content = data.getStringExtra("SCAN_RESULT")
         val uri = Uri.parse(content)
         val fullName = uri.getQueryParameter("iam").toString()
         val githubID = uri.getQueryParameter("gh").toString()
@@ -73,18 +48,19 @@ class FriendListViewModel : ViewModel() {
             twitterID = twitterID
         )
         scope.launch(Dispatchers.IO) {
+            Log.d("DojoApp", "$friend をセーブしました")
             repository.insert(friend)
         }
         _state.value = _state.value?.copy(friendList = repository.allFriends.value)
     }
 
-    fun getQRReadResult(context: Context, requestCode: Int, resultCode: Int, data: Intent?) {
+    fun getQRReadResult(requestCode: Int, resultCode: Int, data: Intent?) {
         if (requestCode == 0 && resultCode == Activity.RESULT_OK && data != null) {
+            Log.d("DojoApp", "QRコード結果取得成功")
             _state.value = _state.value?.copy(isReadSuccess = true, readResultData = data)
-            Toast.makeText(context, data.toString(), Toast.LENGTH_SHORT).show()
-        } else {
+        } else if (resultCode == Activity.RESULT_CANCELED) {
+            Log.d("DojoApp", "QRコード結果取得失敗")
             _state.value = _state.value?.copy(isReadSuccess = false, readResultData = null)
-            Toast.makeText(context, "読み取り失敗", Toast.LENGTH_SHORT).show()
         }
     }
 
@@ -92,29 +68,4 @@ class FriendListViewModel : ViewModel() {
         super.onCleared()
         job.cancel()
     }
-//    fun fetchFriendList(context: Context, owner: LifecycleOwner) {
-//        val db = Room.databaseBuilder(context, MyDatabase::class.java, "user").build()
-//        db.userDao().getUsers().observe(owner, Observer {
-//            _state.value = _state.value?.copy(friendList = it)
-//        })
-//    }
-
-//    private fun saveFriend(context: Context, data: Intent?) {
-//        val content = data?.getStringExtra("SCAN_RESULT")
-//        val uri = Uri.parse(content)
-//        val fullName = uri.getQueryParameter("iam").toString()
-//        val githubID = uri.getQueryParameter("gh").toString()
-//        val twitterID = uri.getQueryParameter("tw").toString()
-//
-//        val db = Room.databaseBuilder(context, MyDatabase::class.java, "user").build()
-//        val user = User(
-//            fullName = fullName,
-//            gitHub = githubID,
-//            twitter = twitterID
-//        )
-//
-//        viewModelScope.launch {
-//            db.userDao().insert(user)
-//        }
-//    }
 }
